refactor: :hammer: change structure
