+++ raqin-api/app/book/service.go
