--- raqin-api/core/model/book.go
@@ -1,10 +1,0 @@
-package model
-
-import "os"
-
-type Book struct {
-	Name   string
-	Author string
-	File   os.File
-	Path   string
-}
